requirements.txt
aiogram==3.5.0
aiogram_dialog==2.1.0
environs==10.0.0
SQLAlchemy~=2.0
asyncpg==0.28.0
apscheduler==3.10.4
pillow==10.3.0
moviepy==1.0.3
numpy==1.26.4
redis